feat: add test deployments command
adds a test-deployments command to the deploy command group which will
generate a deployment, deploymentest and then run the tests against the
generated outputs

refactors the deploy command group into indvidual files as it was
getting a bit busy
