+++ hamlet-cli/hamlet/command/deploy/test.py
@@ -1,0 +1,143 @@
+import os
+import tempfile
+import click
+
+from hamlet.command.common import exceptions
+from hamlet.command.common.config import pass_options
+from hamlet.command.common.exceptions import CommandError
+from hamlet.backend.create import template as create_template_backend
+from hamlet.backend.manage import stack as manage_stack_backend
+from hamlet.backend.manage import deployment as manage_deployment_backend
+from hamlet.backend.test.generate import run as test_generate_backend
+from hamlet.backend.test import run as test_run_backend
+
+from .util import find_deployments_from_options
+
+@click.command(
+    'test-deployments',
+    short_help='',
+    context_settings=dict(
+        max_content_width=240
+    )
+)
+@click.option(
+    '-m',
+    '--deployment-mode',
+    default='update',
+    help='The deployment mode to use for the deployment'
+)
+@click.option(
+    '-l',
+    '--deployment-group',
+    default='.*',
+    show_default=True,
+    help='The deployment group pattern to match',
+)
+@click.option(
+    '-u',
+    '--deployment-unit',
+    default=['.*'],
+    show_default=True,
+    multiple=True,
+    help='The deployment unit pattern to match'
+)
+@click.option(
+    '-o',
+    '--output-dir',
+    type=click.Path(
+        file_okay=False,
+        dir_okay=True,
+        writable=True,
+        readable=True,
+    ),
+    help='The directory where tests are run - temp dir by default'
+)
+@click.option(
+    '-p',
+    '--pytest-args',
+    'pytest_args',
+    help='additional arguments for pytest'
+)
+@click.option(
+    '-s',
+    '--silent',
+    'silent',
+    is_flag=True,
+    help='minimize pytest output'
+)
+@exceptions.backend_handler()
+@pass_options
+def test_deployments(
+        options,
+        deployment_mode,
+        deployment_group,
+        deployment_unit,
+        output_dir,
+        pytest_args,
+        silent,
+        **kwargs):
+    """
+    Create deployments and test them using their test cases
+    """
+    temp_dir = None
+
+    if output_dir is None:
+        temp_dir = tempfile.TemporaryDirectory()
+        output_dir = temp_dir.name
+
+    output_dir = os.path.abspath(output_dir)
+
+    deployments = find_deployments_from_options(
+                    options=options,
+                    deployment_mode=deployment_mode,
+                    deployment_group=deployment_group,
+                    deployment_units=deployment_unit
+                )
+
+    if len(deployments) == 0:
+        raise exceptions.CommandError('No deployments found that match pattern')
+
+    click.echo('')
+    click.echo(click.style(f'[*] Creating deployments:', bold=True, fg='green'))
+    for deployment in deployments:
+
+        deployment_group = deployment['DeploymentGroup']
+        deployment_unit = deployment['DeploymentUnit']
+
+        click.echo(f'[-] {deployment_group}/{deployment_unit}')
+
+        generate_args = {
+            **options.opts,
+            'entrance': 'deployment',
+            'deployment_group': deployment_group,
+            'deployment_unit': deployment_unit,
+            'output_dir': output_dir,
+        }
+        create_template_backend.run(**generate_args, _is_cli=False)
+
+        generate_args = {
+            **options.opts,
+            'entrance': 'deploymenttest',
+            'deployment_group': deployment_group,
+            'deployment_unit': deployment_unit,
+            'output_dir': output_dir,
+        }
+        create_template_backend.run(**generate_args, _is_cli=False)
+
+    click.echo('')
+    click.echo(click.style('[*] Testing deployments:', bold=True, fg='green'))
+    test_script_filename = f'test_deployments.py'
+    test_script = test_generate_backend(
+        output=f'{output_dir}/{test_script_filename}',
+        directory=output_dir,
+    )
+
+    try:
+        test_run_backend.run(
+            testpaths=[ test_script_filename ],
+            pytest_args=pytest_args,
+            silent=silent,
+            root_dir=output_dir
+        )
+    except Exception as e:
+        raise CommandError(str(e))
